+++ src/main/java/com/sieng/java/phoneshop_sieng/service/BrandService.java
@@ -10,8 +10,8 @@ import com.sieng.java.phoneshop_sieng.entity.Brand;
 public interface BrandService {
 	
 	Brand create(Brand brand);
-	Brand getbyId(Integer id);
-	Brand update(Integer id ,Brand brandUpdate);
+	Brand getbyId(Long id);
+	Brand update(Long id ,Brand brandUpdate);
 	List<Brand> getBrands(String name);
 	//List<Brand> getBrands(Map<String, String>params);
 	Page<Brand> getBrands(Map<String, String>params);
